Move stubs in harness to stubs folder
