+++ verification/cbmc/proofs/aws_cryptosdk_priv_try_gen_key/aws_cryptosdk_priv_try_gen_key_harness.c
@@ -17,7 +17,6 @@
  */
 
 #include <aws/cryptosdk/materials.h>
-#include <aws/cryptosdk/private/keyring_trace.h>
 #include <cbmc_invariants.h>
 #include <cipher_openssl.h>
 #include <make_common_data_structures.h>
@@ -27,153 +26,17 @@
 #include <proof_helpers/utils.h>
 
 #include <aws/cryptosdk/private/cipher.h>
-#include <aws/cryptosdk/private/hkdf.h>
 #include <aws/cryptosdk/private/session.h>
 #include <aws/cryptosdk/session.h>
 
-/* Stub this for performance but check the preconditions.
- No modified data structure is used again.
- Original function is here:
- https://github.com/aws/aws-encryption-sdk-c/blob/master/source/hkdf.c#148 */
-int aws_cryptosdk_hkdf(
-    struct aws_byte_buf *okm,
-    enum aws_cryptosdk_sha_version which_sha,
-    const struct aws_byte_buf *salt,
-    const struct aws_byte_buf *ikm,
-    const struct aws_byte_buf *info) {
-    assert(aws_byte_buf_is_valid(salt));
-    assert(aws_byte_buf_is_valid(ikm));
-    assert(aws_byte_buf_is_valid(info));
-    if (nondet_bool()) {
-        return AWS_OP_ERR;
-    }
-    return AWS_OP_SUCCESS;
-}
-
-/* Stub this because of the override of aws_array_list_get_at_ptr and for performance.
- The contents of session->keyring_trace are nondet in the construction of the harness.
- Also neither keyring trace is later used.
- Original function is here:
- https://github.com/aws/aws-encryption-sdk-c/blob/master/source/list_utils.c#L17 */
-int aws_cryptosdk_transfer_list(struct aws_array_list *dest, struct aws_array_list *src) {
-    assert(src != dest);
-    assert(aws_array_list_is_valid(dest));
-    assert(aws_array_list_is_valid(src));
-    assert(dest->item_size == src->item_size);
-
-    if (nondet_bool()) {
-        return AWS_OP_ERR;
-    }
-    return AWS_OP_SUCCESS;
-}
-
-void array_list_item_generator(struct aws_array_list *elems) {
-    assert(elems->item_size == sizeof(struct aws_hash_element));
-    for (size_t index = 0; index < elems->length; ++index) {
-        struct aws_hash_element *val = (struct aws_hash_element *)((uint8_t *)elems->data + (elems->item_size * index));
-        /* Due to the cast to uint16, the entire size of the enc_ctx must be less than < UINT16_MAX
-         * This is a simple way to ensure this without a call to enc_ctx_size. */
-        struct aws_string *key = ensure_string_is_allocated_nondet_length();
-        __CPROVER_assume(aws_string_is_valid(key));
-        __CPROVER_assume(key->len <= UINT8_MAX);
-        val->key                 = key;
-        struct aws_string *value = ensure_string_is_allocated_nondet_length();
-        __CPROVER_assume(aws_string_is_valid(value));
-        __CPROVER_assume(value->len <= UINT8_MAX);
-        val->value = value;
-    }
-}
-
-/* Stub this because of the override of aws_array_list_get_at_ptr
- Original function is here:
- https://github.com/aws/aws-encryption-sdk-c/blob/master/source/keyring_trace.c#L235 */
-void aws_cryptosdk_keyring_trace_clear(struct aws_array_list *trace) {
-    AWS_FATAL_PRECONDITION(aws_cryptosdk_keyring_trace_is_valid(trace));
-    AWS_FATAL_PRECONDITION(trace->item_size == sizeof(struct aws_cryptosdk_keyring_trace_record));
-    aws_array_list_clear(trace);
-}
-
-/* Stub this until https://github.com/diffblue/cbmc/issues/5344 is fixed
- Original function is here:
- https://github.com/aws/aws-encryption-sdk-c/blob/master/source/edk.c#L44 */
-void aws_cryptosdk_edk_list_clean_up(struct aws_array_list *encrypted_data_keys) {
-    assert(aws_cryptosdk_edk_list_is_valid(encrypted_data_keys));
-    aws_array_list_clean_up(encrypted_data_keys);
-}
-
-/**
- * Receives encryption request from user and attempts to generate encryption materials,
- * including an encrypted data key and a list of EDKs for doing encryption.
- *
- * On success returns AWS_OP_SUCCESS and allocates encryption materials object at address
- * pointed to by output.
- *
- * On failure returns AWS_OP_ERR, sets address pointed to by output to NULL, and sets
- * internal AWS error code.
- */
 int generate_enc_materials(
     struct aws_cryptosdk_cmm *cmm,
     struct aws_cryptosdk_enc_materials **output,
-    struct aws_cryptosdk_enc_request *request) {
-    assert(aws_cryptosdk_cmm_base_is_valid(cmm));
-    assert(AWS_OBJECT_PTR_IS_WRITABLE(output));
-    assert(aws_cryptosdk_enc_request_is_valid(request));
-
-    struct aws_cryptosdk_enc_materials *materials = malloc(sizeof(*materials));
-    if (materials == NULL) {
-        *output = NULL;
-        return AWS_OP_ERR;
-    }
-
-    // Set up the allocator
-    // Request->alloc is session->alloc
-    materials->alloc = request->alloc;
-    __CPROVER_assume(aws_allocator_is_valid(materials->alloc));
-    __CPROVER_assume(materials->alloc != NULL);
-
-    // Set up the signctx
-    materials->signctx = ensure_nondet_sig_ctx_has_allocated_members();
-    __CPROVER_assume(aws_cryptosdk_sig_ctx_is_valid_cbmc(materials->signctx));
-
-    // Set up the unencrypted_data_key
-    __CPROVER_assume(aws_byte_buf_is_bounded(&materials->unencrypted_data_key, MAX_BUFFER_SIZE));
-    ensure_byte_buf_has_allocated_buffer_member(&materials->unencrypted_data_key);
-    __CPROVER_assume(aws_byte_buf_is_valid(&materials->unencrypted_data_key));
-
-    // Set up the edk_list
-    __CPROVER_assume(aws_cryptosdk_edk_list_is_bounded(&materials->encrypted_data_keys, MAX_EDK_LIST_ITEMS));
-    ensure_cryptosdk_edk_list_has_allocated_list(&materials->encrypted_data_keys);
-    __CPROVER_assume(aws_cryptosdk_edk_list_is_valid(&materials->encrypted_data_keys));
-
-    // The alloc is session->alloc
-    materials->encrypted_data_keys.alloc = request->alloc;
-
-    // Precondition: The edk_list has valid list elements
-    __CPROVER_assume(aws_cryptosdk_edk_list_elements_are_bounded(&materials->encrypted_data_keys, MAX_BUFFER_SIZE));
-    ensure_cryptosdk_edk_list_has_allocated_list_elements(&materials->encrypted_data_keys);
-    __CPROVER_assume(aws_cryptosdk_edk_list_elements_are_valid(&materials->encrypted_data_keys));
-
-    // Set up the keyring trace
-    __CPROVER_assume(aws_array_list_is_bounded(
-        &materials->keyring_trace, MAX_TRACE_LIST_ITEMS, sizeof(struct aws_cryptosdk_keyring_trace_record)));
-    __CPROVER_assume(materials->keyring_trace.item_size == sizeof(struct aws_cryptosdk_keyring_trace_record));
-    ensure_array_list_has_allocated_data_member(&materials->keyring_trace);
-    __CPROVER_assume(aws_array_list_is_valid(&materials->keyring_trace));
-    ensure_trace_has_allocated_records(&materials->keyring_trace, MAX_STRING_LEN);
-    __CPROVER_assume(aws_cryptosdk_keyring_trace_is_valid(&materials->keyring_trace));
-
-    // Set the alg_id
-    enum aws_cryptosdk_alg_id alg_id;
-    materials->alg = alg_id;
-
-    *output = materials;
-    return AWS_OP_SUCCESS;
-}
+    struct aws_cryptosdk_enc_request *request);
 
 void aws_cryptosdk_priv_try_gen_key_harness() {
     /* Nondet Input */
-    struct aws_cryptosdk_session *session =
-        session_setup(MAX_TABLE_SIZE, MAX_TRACE_LIST_ITEMS, MAX_EDK_LIST_ITEMS, MAX_BUFFER_SIZE, MAX_STRING_LEN);
+    struct aws_cryptosdk_session *session = malloc(sizeof(*session));
 
     /* Assumptions */
     __CPROVER_assume(session != NULL);
